jfk-thesis-main-2/rag/requirements.txt
flask
flask-cors
duckdb
groq
python-dotenv
pandas
